Add separate tables for each data type
Rather than store everything in a single blob, create tables for each
data type. Each has a foreign key to reference it's parent in a simple
many-to-one relationship.

This allows the UI to only pass the path and the current id around. That
ID is used to lookup the resource being edited, and allows the UI to
easily reference the parent (through the foreign key) for the Back
button.

Queries will need to be written to get the children of the resource for
the lists on each page.