Persist Remote Data
Add in API routes for uploading and downloading remote data through
firebase. Firebase was choosen because it was a serverless nosql
database without schema restraints allowing for quick iteration.

The database does not need to be Firebase in future, it can be swapped
out for something more traditional if choosen. Since the development app
is hosted on `vercel` though, another serverless database was ideal.

Add in a new `/data` page for looking at the uploaded results and
downloading them. The data can be downloaded as CSV. For this iteration,
the data is not in the best format, it's mostly just a JSON blob, but we
can update it once we know how many CSV files to create and download.

Upload button was enabled, but a better UX needs to be present around
uploading.